separate serialization check from properties check in ModelTest
